% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSetting.R
\name{AcceptTermsAndConditions}
\alias{AcceptTermsAndConditions}
\title{Accept terms and conditions, only run once per session, must agree to terms or functions will not work}
\usage{
AcceptTermsAndConditions(accepted = FALSE)
}
\arguments{
\item{accepted}{Default=FALSE , BOOLEAN}
}
\description{
Accept terms and conditions, only run once per session, must agree to terms or functions will not work
}
\examples{
AcceptTermsAndConditions(accepted=TRUE)
}

