% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authorization.R
\name{authorization}
\alias{authorization}
\title{Authorization for the Pinnacle API}
\usage{
authorization(user, pwd)
}
\arguments{
\item{user}{Pinnacle Username}

\item{pwd}{Pinnacle Password}
}
\description{
Authorization for the Pinnacle API
}
