% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showOddsFunctions.R
\name{cbindNames}
\alias{cbindNames}
\title{Combines elements of an object, keeping names}
\usage{
cbindNames(x)
}
\arguments{
\item{x}{takes a list of elements with names}
}
\value{
data.frame containing combined names
}
\description{
Combines elements of an object, keeping names
}

