% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSetting.R
\name{CheckTermsAndConditions}
\alias{CheckTermsAndConditions}
\title{Prompts User for Terms and Conditions, otherwise stops running function}
\usage{
CheckTermsAndConditions()
}
\description{
Prompts User for Terms and Conditions, otherwise stops running function
}
\examples{
CheckTermsAndConditions()
}

