% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showOddsFunctions.R
\name{combineFactors}
\alias{combineFactors}
\title{Combines the list at given depth with the factors associated with that depth}
\usage{
combineFactors(x, depth = 4)
}
\arguments{
\item{x}{A list}

\item{depth}{Number of levels of sublists}
}
\value{
a tree with the factors at the defined level combined with a df
}
\description{
Combines the list at given depth with the factors associated with that depth
}

