% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixNames.R
\name{FixNames}
\alias{FixNames}
\title{Fix up names to be shorter}
\usage{
FixNames(dt, lastx = 2)
}
\arguments{
\item{dt}{a data.table with pinnacle.API standard names}

\item{lastx}{last x names to use as part of an identifier}
}
\value{
same data.frame with names fixed to length x indicators
}
\description{
Fix up names to be shorter
}
