% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSetting.R
\name{GetAPIEndpoint}
\alias{GetAPIEndpoint}
\title{Gets the current API endpoint}
\usage{
GetAPIEndpoint()
}
\value{
the currently set API endpoint
}
\description{
Gets the current API endpoint
}
\examples{
SetAPIEndpoint("https://api.pinnaclesports.com/v2/")
GetAPIEndpoint()
SetAPIEndpoint("https://api.pinnaclesports.com")
}
