% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetBets.R
\name{GetBetsList}
\alias{GetBetsList}
\title{Get a list of running/settled bets}
\usage{
GetBetsList(betlist = c("SETTLED", "RUNNING"), fromDate = Sys.Date() - 7,
  toDate = Sys.Date() + 7)
}
\arguments{
\item{betlist}{Either 'SETTLED' or 'RUNNING' Default Behavior shows both}

\item{fromDate}{Iso8061 Date Default: 7 days prior, Sys.Date()-7}

\item{toDate}{Iso8061 Date  Default: Line Close within next 7 days Sys.Date()+7}
}
\value{
A list of bets and associated details
}
\description{
Get a list of running/settled bets
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted=TRUE)
GetBetsList()}
}

