% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetClientBalance.R
\name{GetClientBalance}
\alias{GetClientBalance}
\title{Get Client Balance}
\usage{
GetClientBalance(force = TRUE)
}
\arguments{
\item{force}{Default=TRUE , boolean if TRUE force a reload of the data if FALSE use cached data}
}
\value{
vector client balance parameter
}
\description{
Get Client Balance
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted=TRUE)
GetClientBalance()}
}

