% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EnvironmentSetting.R
\name{GetCredentials}
\alias{GetCredentials}
\title{Get your credential values}
\usage{
GetCredentials()
}
\value{
A data.frame with your username and password
}
\description{
Get your credential values
}
\examples{
SetCredentials("TESTAPI","APITEST")
GetCredentials()
}

