% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCurrencies.R
\name{GetCurrencies}
\alias{GetCurrencies}
\title{Get the list of supported Currencies}
\usage{
GetCurrencies(force = FALSE)
}
\arguments{
\item{force}{Default=FALSE, boolean if TRUE force a reload of the data if FALSE use cached data}
}
\value{
a data frame with these columns:
\itemize{
\item Currency Code
\item Exchange Rate to USD
\item Currency Name
}
}
\description{
Get the list of supported Currencies
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted=TRUE)
GetCurrencies()}
}

