% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetInrunning.R
\name{GetInrunning}
\alias{GetInrunning}
\title{GetInrunning}
\usage{
GetInrunning()
}
\value{
A dataframe containing the current State of live events
}
\description{
GetInrunning
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted=TRUE)
GetInrunning()
}
}

