% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSettledFixtures.R
\name{GetSettledFixtures}
\alias{GetSettledFixtures}
\title{Get Settled Fixtures}
\usage{
GetSettledFixtures(sportid, leagueids = NULL, since = FALSE)
}
\arguments{
\item{sportid}{}

\item{leagueids}{}

\item{since}{}
}
\value{
a data.frame of settled fixtures
}
\description{
Get Settled Fixtures
}

