% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSettledSpecialFixtures.R
\name{GetSettledSpecialFixtures}
\alias{GetSettledSpecialFixtures}
\title{Get Settled Special Fixtures}
\usage{
GetSettledSpecialFixtures(sportid, leagueids = NULL, since = NULL)
}
\arguments{
\item{sportid}{}

\item{leagueids}{}

\item{since}{}
}
\value{
a data.frame of settled special fixtures
}
\description{
Get Settled Special Fixtures
}

