% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpecialFixtures.R
\name{GetSpecialFixtures}
\alias{GetSpecialFixtures}
\title{Get Special Fixtures}
\usage{
GetSpecialFixtures(sportid, leagueids = NULL, category = NULL,
  eventid = NULL, specialid = NULL, since = NULL)
}
\arguments{
\item{sportid}{}

\item{leagueids}{}

\item{category}{}

\item{eventid}{}

\item{specialid}{}

\item{since}{}
}
\value{
a data.frame of settled special fixtures
}
\description{
Get Special Fixtures
}

