% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSports.R
\name{GetSports}
\alias{GetSports}
\title{Get Sports}
\usage{
GetSports(force = FALSE)
}
\arguments{
\item{force}{Default=FALSE, boolean if TRUE force a reload of the data if FALSE use cached data}
}
\value{
a data frame with these columns:
\itemize{
\item SportID
\item LinesAvailable
\item SportName
}
}
\description{
Returns all sports with the status whether they currently have lines or not
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted=TRUE)
GetSports()}
}

