% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeaserOdds.R
\name{GetTeaserOdds}
\alias{GetTeaserOdds}
\title{Get Teaser Odds}
\usage{
GetTeaserOdds(teaserid, force = TRUE)
}
\arguments{
\item{teaserid}{id denoting teaser}

\item{force}{boolean if FALSE, functions using cached data will use the values since the last force}
}
\value{
data.frame of odds
}
\description{
Get Teaser Odds
}
\examples{
\donttest{

}

}

