% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSetting.R
\name{GetUsername}
\alias{GetUsername}
\title{Get your Username}
\usage{
GetUsername()
}
\value{
A String, your current username
}
\description{
Get your Username
}
\examples{
SetCredentials("TESTAPI","APITEST")
GetUsername()
}

