% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showOddsFunctions.R
\name{JSONtoDF}
\alias{JSONtoDF}
\title{Sets the determined depth as a data.frame, and evades R's list-dropping rules}
\usage{
JSONtoDF(x, depth = 5, flag = TRUE)
}
\arguments{
\item{x}{a list}

\item{depth}{numeric value of the number of sublist}

\item{flag}{whether to start with combining factors or binding rows}
}
\value{
a tree with the defined level as a dataframe
}
\description{
Sets the determined depth as a data.frame, and evades R's list-dropping rules
}

