% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnRawXMLResponse.R
\name{pinnRawXMLResponse}
\alias{pinnRawXMLResponse}
\title{pinnRawXMLResponse}
\usage{
pinnRawXMLResponse()
}
\value{
a list containing the raw and the converted XML responses
}
\description{
pinnRawXMLResponse
}

