% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showOddsFunctions.R
\name{replaceNulls}
\alias{replaceNulls}
\title{Set all null fields to NA}
\usage{
replaceNulls(x)
}
\arguments{
\item{x}{an R object}
}
\value{
the same object with nulls replaced
}
\description{
Set all null fields to NA
}
\examples{
replaceNulls(list(1,NULL, list(1, NULL, list(1,NULL))))
}

