% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSetting.R
\name{SetAPIEndpoint}
\alias{SetAPIEndpoint}
\title{Sets the API endpoint to use}
\usage{
SetAPIEndpoint(url = "https://api.pinnaclesports.com")
}
\arguments{
\item{url}{a url, default value is the usual API endpoint}
}
\value{
void
}
\description{
Sets the API endpoint to use
}

