% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSetting.R
\name{SetCredentials}
\alias{SetCredentials}
\title{Set your pinnaclesports.com user credentials}
\usage{
SetCredentials(username, password)
}
\arguments{
\item{username}{Your username}

\item{password}{Your password}
}
\description{
Set your pinnaclesports.com user credentials
}
\examples{
SetCredentials("TESTAPI","APITEST")
}
