% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showOddsDF.R
\name{showOddsDF}
\alias{showOddsDF}
\title{Combine Odds, Fixture, and In-Running Information into a Single Data Frame}
\usage{
showOddsDF(sportid, leagueids = NULL, eventids = NULL, since = NULL,
  islive = 0, force = TRUE, tableformat = "mainlines", namesLength = 3,
  attachLeagueInfo = TRUE, oddsformat = "AMERICAN", fixtures_since = NULL)
}
\arguments{
\item{sportid}{An integer giving the sport. If this is missing in
interactive mode, a menu of options is presented to the user.}

\item{leagueids}{A vector of league IDs, or \code{NULL}.}

\item{eventids}{A vector of event IDs, or \code{NULL}.}

\item{since}{Used to receive incremental odds updates. See
\code{\link{GetOdds}}.}

\item{islive}{When \code{TRUE}, retrieve only live events.}

\item{force}{Currently ignored.}

\item{tableformat}{The format of the odds. See \code{\link{GetOdds}}.}

\item{namesLength}{The number of identifiers to use in the names.}

\item{attachLeagueInfo}{When \code{TRUE}, include league information in the
data.}

\item{oddsformat}{Format for the returned odds. See \code{\link{GetOdds}}.}

\item{fixtures_since}{Used to receive incremental fixture updates. See
\code{\link{GetFixtures}}.}
}
\value{
A data frame combining odds and fixtures data, containing \code{NA}s where
levels of factors do not have a value. Example of the naming convention:
\code{spread.altLineId.N} is the \code{altLineId} associated with
\code{spread.hdp.(N+1)}, whereas \code{spread.hdp} refers to the mainline.
\code{spread.altLineId} is the first alternate, and equivalent to
\code{spread.altLineId.0}.
}
\description{
Queries the event listing for a given sport and the odds offered on each
event. This query can be filtered by league and/or event ID, and narrowed to
include only live events.
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted=TRUE)
# Run without arguments, it will prompt you for the sport
showOddsDF()}

}
\seealso{
See \code{\link{GetOdds}}, \code{\link{GetFixtures}}, and
\code{\link{GetInrunning}} for the underlying API requests.
}
