% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Authorization.R
\name{authorization}
\alias{authorization}
\title{Authorization for the Pinnacle API}
\usage{
authorization(user = as.character(GetCredentials()$Value[1]),
  pwd = as.character(GetCredentials()$Value[2]))
}
\arguments{
\item{user}{Pinnacle Username}

\item{pwd}{Pinnacle Password}
}
\description{
Authorization for the Pinnacle API
}

