% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetLeagues.R
\name{GetLeagues}
\alias{GetLeagues}
\title{Get Leagues for Sport(s) by name}
\usage{
GetLeagues(sports, force = FALSE, regex = FALSE)
}
\arguments{
\item{sports}{character vector of sports names.}

\item{force}{Default=FALSE, boolean if TRUE force a reload of the data if FALSE use cached data}

\item{regex}{Default=FALSE, boolean if TRUE , retreives sports id using regular expression on names}
}
\value{
a data frame having columns:
\itemize{
\item LeagueID
\item LinesAvailable
\item HomeTeam
\item AllowRoundRobin
\item LeagueName
}
}
\description{
Returns all Leagues for the Sport(s)
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted=TRUE)
GetLeagues("Badminton")}
}

