% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetLine.R
\name{GetLine}
\alias{GetLine}
\title{Get Lines
(Use to get more detail on a single line, but the GetOdds or showOddsDF versions are intended for large amounts of data)}
\usage{
GetLine(sportId, leagueids, eventId, periodNumber, betType, team = NULL,
  side = NULL, handicap = NULL, oddsFormat = "AMERICAN")
}
\arguments{
\item{sportId}{The sport ID}

\item{leagueids}{integer vector of leagueids.}

\item{eventId}{numeric xxxxx}

\item{periodNumber}{xxxxx}

\item{betType}{xxxx}

\item{team}{xxxx}

\item{side}{xxx}

\item{handicap}{xxx}

\item{oddsFormat}{xxx}
}
\value{
returns a data frame with columns:
\itemize{
\item SportID
\item Last
\item League
\item LeagueID
\item EventID
\item StartTime
\item HomeTeamName
\item AwayTeamName
\item Rotation Number
\item Live Status
\item Status
\item Parlay Status
}
}
\description{
Get Lines
(Use to get more detail on a single line, but the GetOdds or showOddsDF versions are intended for large amounts of data)
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted=TRUE)
 GetLine(sportId=1,leagueids=191545,eventId=495418854,
 periodNumber=0,team="TEAM1",betType="Moneyline")}
}

