% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOdds.R
\name{GetOdds}
\alias{GetOdds}
\title{Get Odds for Non-Settled Events in a Given Sport}
\usage{
GetOdds(sportid, leagueids = NULL, eventids = NULL, since = NULL,
  islive = 0, oddsformat = "AMERICAN", tableformat = "mainlines",
  force = TRUE)
}
\arguments{
\item{sportid}{An integer giving the sport. If this is missing in
interactive mode, a menu of options is presented to the user.}

\item{leagueids}{A vector of league IDs, or \code{NULL}.}

\item{eventids}{A vector of event IDs, or \code{NULL}.}

\item{since}{To receive only listings updated since the last query, set
\code{since} to the value of \code{last} from the previous fixtures
response. Otherwise it will query all listings.}

\item{islive}{When \code{TRUE}, retrieve only live events.}

\item{oddsformat}{Format for the returned odds. One of \code{"AMERICAN"},
\code{"DECIMAL"}, \code{"HONGKONG"}, \code{"INDONESIAN"}, or
\code{"MALAY"}.}

\item{tableformat}{One of
\itemize{
  \item \code{"mainlines"} for mainlines only (the default);
  \item \code{"long"} for a single record for each spread/total on an event; 
  \item \code{"wide"} for all lines as one record; or
  \item \code{"subtables"} all lines for spreads/totals stored as nested
    tables.
}}

\item{force}{Currently ignored.}
}
\value{
A data frame of odds.
}
\description{
Queries all odds offered for a given sport, which can be filtered by league
and/or event ID, and narrowed to include only live events.
}
\details{
This function will raise an error if the API does not return HTTP status
\code{OK}. For information on the possible errors, see the API documentation
for \href{https://pinnacleapi.github.io/#operation/Odds_Straight_V1_Get}{Get Odds}.
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted = TRUE)
# We can run without parameters, and will be given a selection of sports
GetOdds()}

}
