% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetOdds.R
\name{GetOdds}
\alias{GetOdds}
\title{Get Odds}
\usage{
GetOdds(sportname, leagueIds, since = NULL, isLive = 0)
}
\arguments{
\item{sportname}{The sport name for which to retrieve the fixutres}

\item{leagueIds}{integer vector of leagueids.}

\item{since}{numeric This is used to receive incremental updates.
Use the value of last from previous fixtures response.}

\item{isLive}{boolean if TRUE retrieves ONLY live events}
}
\value{
list of lists
}
\description{
Get Odds
}
\examples{
\donttest{
SetCredentials("TESTAPI","APITEST")
AcceptTermsAndConditions(accepted=TRUE)
GetOdds (sportname="Badminton", leagueIds=191545,,isLive=0)}
}

