% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSettledSpecialFixtures.R
\name{GetSettledSpecialFixtures}
\alias{GetSettledSpecialFixtures}
\title{Get Settled Special Fixtures}
\usage{
GetSettledSpecialFixtures(sportid, leagueids = NULL, since = NULL)
}
\arguments{
\item{sportid}{(optional) an integer giving the sport, if missing, a menu of options is presented}

\item{leagueids}{(optional) integer vector with league IDs.}

\item{since}{(optional) numeric This is used to receive incremental updates.
Use the value of last from previous fixtures response.}
}
\value{
a data.frame of settled special fixtures
}
\description{
Get Settled Special Fixtures
}
\examples{
\donttest{
SetCredentials("TESTAPI", "APITEST")
AcceptTermsAndConditions(accepted=TRUE)
# Can be run without arguments
GetSettledSpecialFixtures()}
}

