% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpecialFixtures.R
\name{GetSpecialFixtures}
\alias{GetSpecialFixtures}
\title{Get Special Fixtures}
\usage{
GetSpecialFixtures(sportid, leagueids = NULL, category = NULL,
  eventid = NULL, specialid = NULL, since = NULL)
}
\arguments{
\item{sportid}{(optional) an integer giving the sport, if missing, a menu of options is presented}

\item{leagueids}{(optional) integer vector with league IDs.}

\item{category}{(optional) See API Manual}

\item{eventid}{(optional) Associated event ID}

\item{specialid}{(optional) Associated special ID}

\item{since}{(optional) numeric This is used to receive incremental updates.
Use the value of last from previous fixtures response.}
}
\value{
a data.frame of special fixtures
}
\description{
Get Special Fixtures
}
\examples{
\donttest{
SetCredentials("TESTAPI", "APITEST")
AcceptTermsAndConditions(accepted=TRUE)
GetSpecialFixtures()}
}

